ctdb-protocol: add CTDB_REC_FLAG_PERSISTENT[_IN_PROGRESS]
CTDB_REC_FLAG_PERSISTENT is set when a record was stored once with per-record
persistency semantics and has a record in the persistent backup database. It's
used by the dbwrap_ctdb backend to determine if the backup record can be deleted
when storing a record.

CTDB_REC_FLAG_PERSISTENT_IN_PROGRESS is used by dbwrap_ctdb when opening database with per-record
persistency support (DBWRAP_FLAG_PER_REC_PERSISTENT). The flag is needed to make
stores atomic without holding a TDB chainlock, similar to a "glock" in smbd
parlance and also comes with a "glock" owner payload (struct server_id) to
detect stale "glocks".

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Martin Schwenke <martin@meltin.net>