scavenger: pass individual args not fsp to scavenger_schedule_disconnected()
In preperation of calling scavenger_schedule_disconnected() from a
function where we don't have an fsp.

Not having an fsp also means not having access to the open record via
fsp->op->global to fetch the timeout values. Hence in the scavenger in
scavenger_add_timer() we use smbXsrv_open_global_lookup() to fetch the open
record with the timeout values.

No change in behaviour.

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Anoop C S <anoopcs@samba.org>