vfs_default: set cookie.allow_reconnect for PH already in vfs_default_durable_cookie()
For Durable Handles we set cookie.allow_reconnect in the SMB_VFS_DISCONNECT()
implemtation, but for Persistent Handles we can't rely on SMB_VFS_DISCONNECT()
being called in case of a server side process or node failure.

The whole disconnect/reconnect logic is therefor adjusted to make
SMB_VFS_DISCONNECT() a no-op for Persistent Handles.

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Anoop C S <anoopcs@samba.org>