ctdb-common: Encapsulate whole IPv6 sendto() address setup
Variable sall is no longer used by IPv4.  It is now only used to
construct the sendto() destination address for IPv6 NA.

Replace this variable with a sockaddr_storage so it can fit larger
link-level addresses.  Change the type of ip6_ll_multicast_build() and
use it to set the sockaddr_storage (as a sockaddr_ll), based on the
link-level broadcast address and RFCs.

IPoIB is currently not supported.

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Reviewed-by: Vinit Agnihotri <vagnihot@redhat.com>
Reviewed-by: Amitay Isaacs <amitay@gmail.com>